requirements.txt
Flask
python-dotenv
Werkzeug
Flask-WTF
Flask-Limiter
gunicorn
